myheroagency-7 Save and load

+++ core/src/com/nuclearthinking/myheroagency/model/Player.java
@@ -2,7 +2,7 @@ package com.nuclearthinking.myheroagency.model;
 
 class Player {
 
-    int level;
+    byte level;
     int exp;
 
 }
